+++ Octokit.Reactive/Clients/IObservableIssuesEventsClient.cs
@@ -20,7 +20,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
-        IObservable<EventInfo> GetAllForIssue(string owner, string name, int number);
+        IObservable<EventInfo> GetAllForIssue(string owner, string name, long number);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -30,7 +30,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
-        IObservable<EventInfo> GetAllForIssue(long repositoryId, int number);
+        IObservable<EventInfo> GetAllForIssue(long repositoryId, long number);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -42,7 +42,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<EventInfo> GetAllForIssue(string owner, string name, int number, ApiOptions options);
+        IObservable<EventInfo> GetAllForIssue(string owner, string name, long number, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -53,7 +53,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<EventInfo> GetAllForIssue(long repositoryId, int number, ApiOptions options);
+        IObservable<EventInfo> GetAllForIssue(long repositoryId, long number, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the repository.
@@ -106,7 +106,7 @@ namespace Octokit.Reactive
         /// <param name="number">The event id</param>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
         Justification = "Method makes a network request")]
-        IObservable<IssueEvent> Get(string owner, string name, int number);
+        IObservable<IssueEvent> Get(string owner, string name, long number);
 
         /// <summary>
         /// Gets a single event
@@ -118,6 +118,6 @@ namespace Octokit.Reactive
         /// <param name="number">The event id</param>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
         Justification = "Method makes a network request")]
-        IObservable<IssueEvent> Get(long repositoryId, int number);
+        IObservable<IssueEvent> Get(long repositoryId, long number);
     }
 }
