+++ Octokit.Reactive/Clients/ObservableIssuesEventsClient.cs
@@ -32,7 +32,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
-        public IObservable<EventInfo> GetAllForIssue(string owner, string name, int number)
+        public IObservable<EventInfo> GetAllForIssue(string owner, string name, long number)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -48,7 +48,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
-        public IObservable<EventInfo> GetAllForIssue(long repositoryId, int number)
+        public IObservable<EventInfo> GetAllForIssue(long repositoryId, long number)
         {
             return GetAllForIssue(repositoryId, number, ApiOptions.None);
         }
@@ -63,7 +63,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        public IObservable<EventInfo> GetAllForIssue(string owner, string name, int number, ApiOptions options)
+        public IObservable<EventInfo> GetAllForIssue(string owner, string name, long number, ApiOptions options)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -81,7 +81,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        public IObservable<EventInfo> GetAllForIssue(long repositoryId, int number, ApiOptions options)
+        public IObservable<EventInfo> GetAllForIssue(long repositoryId, long number, ApiOptions options)
         {
             Ensure.ArgumentNotNull(options, nameof(options));
 
@@ -158,7 +158,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The event id</param>
-        public IObservable<IssueEvent> Get(string owner, string name, int number)
+        public IObservable<IssueEvent> Get(string owner, string name, long number)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -174,7 +174,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The event id</param>
-        public IObservable<IssueEvent> Get(long repositoryId, int number)
+        public IObservable<IssueEvent> Get(long repositoryId, long number)
         {
             return _client.Get(repositoryId, number).ToObservable();
         }
