+++ Octokit/Clients/IIssuesEventsClient.cs
@@ -21,7 +21,7 @@ namespace Octokit
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
-        Task<IReadOnlyList<EventInfo>> GetAllForIssue(string owner, string name, int number);
+        Task<IReadOnlyList<EventInfo>> GetAllForIssue(string owner, string name, long number);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -31,7 +31,7 @@ namespace Octokit
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
-        Task<IReadOnlyList<EventInfo>> GetAllForIssue(long repositoryId, int number);
+        Task<IReadOnlyList<EventInfo>> GetAllForIssue(long repositoryId, long number);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -43,7 +43,7 @@ namespace Octokit
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        Task<IReadOnlyList<EventInfo>> GetAllForIssue(string owner, string name, int number, ApiOptions options);
+        Task<IReadOnlyList<EventInfo>> GetAllForIssue(string owner, string name, long number, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the issue.
@@ -54,7 +54,7 @@ namespace Octokit
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        Task<IReadOnlyList<EventInfo>> GetAllForIssue(long repositoryId, int number, ApiOptions options);
+        Task<IReadOnlyList<EventInfo>> GetAllForIssue(long repositoryId, long number, ApiOptions options);
 
         /// <summary>
         /// Gets all events for the repository.
@@ -107,7 +107,7 @@ namespace Octokit
         /// <param name="number">The event id</param>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
              Justification = "Method makes a network request")]
-        Task<IssueEvent> Get(string owner, string name, int number);
+        Task<IssueEvent> Get(string owner, string name, long number);
 
         /// <summary>
         /// Gets a single event
@@ -119,6 +119,6 @@ namespace Octokit
         /// <param name="number">The event id</param>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
              Justification = "Method makes a network request")]
-        Task<IssueEvent> Get(long repositoryId, int number);
+        Task<IssueEvent> Get(long repositoryId, long number);
     }
 }
